reorganize model folder
