+++ src/layoutparser/models/detectron2/catalog.py
